disentangle configuration for different workers and beats.
