+++ backend/danswer/background/celery/tasks/shared/RedisFenceData.py
@@ -1,0 +1,11 @@
+from datetime import datetime
+
+from pydantic import BaseModel
+
+
+class RedisFenceData(BaseModel):
+    index_attempt_id: int
+    num_tasks: int
+    started: datetime | None
+    submitted: datetime
+    task_id: str
