+++ backend/supervisord.conf
@@ -15,7 +15,7 @@ logfile=/var/log/supervisord.log
 # relatively compute-light (e.g. they tend to just make a bunch of requests to 
 # Vespa / Postgres)
 [program:celery_worker_primary]
-command=celery -A danswer.background.celery.celery_run:celery_app worker 
+command=celery -A danswer.background.celery.versioned_apps.primary worker 
     --pool=threads
     --concurrency=4
     --prefetch-multiplier=1 
@@ -33,7 +33,7 @@ stopasgroup=true
 # since this is often the bottleneck for "sync" jobs (e.g. document set syncing, 
 # user group syncing, deletion, etc.) 
 [program:celery_worker_light]
-command=bash -c "celery -A danswer.background.celery.celery_run:celery_app worker \
+command=bash -c "celery -A danswer.background.celery.versioned_apps.light worker \
     --pool=threads \
     --concurrency=${CELERY_WORKER_LIGHT_CONCURRENCY:-24} \
     --prefetch-multiplier=${CELERY_WORKER_LIGHT_PREFETCH_MULTIPLIER:-8} \
@@ -48,7 +48,7 @@ startsecs=10
 stopasgroup=true
 
 [program:celery_worker_heavy]
-command=celery -A danswer.background.celery.celery_run:celery_app worker 
+command=celery -A danswer.background.celery.versioned_apps.heavy worker 
     --pool=threads
     --concurrency=4
     --prefetch-multiplier=1 
@@ -63,7 +63,7 @@ startsecs=10
 stopasgroup=true
 
 [program:celery_worker_indexing]
-command=bash -c "celery -A danswer.background.celery.celery_run:celery_app worker \
+command=bash -c "celery -A danswer.background.celery.versioned_apps.indexing worker \
     --pool=threads \
     --concurrency=${CELERY_WORKER_INDEXING_CONCURRENCY:-1} \
     --prefetch-multiplier=1 \
@@ -79,7 +79,7 @@ stopasgroup=true
 
 # Job scheduler for periodic tasks
 [program:celery_beat]
-command=celery -A danswer.background.celery.celery_run:celery_app beat
+command=celery -A danswer.background.celery.versioned_apps.beat beat
 stdout_logfile=/var/log/celery_beat.log
 stdout_logfile_maxbytes=16MB
 redirect_stderr=true
